requirements.txt
streamlit
python-dotenv

langchain-core
langchain-community
langchain-text-splitters
langchain-openai
langchain-oracledb

pypdf
oracledb
